Move testAccPreCheckHelmRepositoryDestroy() to resource_release_test.go
